Fixing postgres host config
This is a mismatch between the pg driver and what comes out of the
url.parse() function. The pg driver expects host to not have a port
number while url.parse() returns hostname as not having a port number.
The pg driver performs a similar translation internally when using a
raw string as the connection config.